requirements.txt
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
h11==0.14.0
idna==3.4
Js2Py==0.71
outcome==1.2.0
Pillow==9.2.0
pycparser==2.21
pycryptodome==3.15.0
pyjsparser==2.7.1
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
readchar==4.0.3
regex==2022.9.13
requests==2.28.1
secure-smtplib==0.1.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tkcalendar==1.6.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
wsproto==1.2.0
